Now I can handle the events without doubling them
