+++ src/services/tracker/utils.ts
@@ -1,3 +1,4 @@
+import { TextEditor } from "vscode";
 import { LocalStorage } from "../data/LocalStorage";
 
 const isStoraged = (fileName: string, storageManager: LocalStorage) => {
@@ -7,4 +8,15 @@ const isStoraged = (fileName: string, storageManager: LocalStorage) => {
   })
 }
 
-export { isStoraged }
+const addRecord = (instance: TextEditor, storageManager: LocalStorage) => {
+  const record = {
+    languageId: instance.document.languageId,
+    fileName: instance.document.fileName,
+    time: 0
+  }
+  const currentUserTime: any[] = storageManager.getValue("GHUserTime")
+  currentUserTime.push(record)
+  storageManager.setValue("GHUserTime", currentUserTime)
+}
+
+export { isStoraged, addRecord }
